requirements.txt
# Recommended Python: 3.10-3.13
# Core Dependencies
python-dotenv==1.0.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
google-genai==1.7.0

# LangChain
langchain==0.1.0
langchain-community==0.0.13
langchain-text-splitters==0.0.1
langchain-google-genai==0.0.6

# Embeddings & Vector Store
sentence-transformers==2.2.2
faiss-cpu==1.7.4
# For GPU: faiss-gpu==1.7.4

# ML/AI
torch==2.1.2
transformers==4.36.2

# Utilities
numpy==1.24.3
tqdm==4.66.1
pytest==8.3.5
